requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asttokens==2.4.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.7
decorator==5.1.1
distro==1.9.0
executing==2.1.0
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
idna==3.10
ipython==8.29.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.7.1
jsonpatch==1.33
jsonpickle==4.0.0
jsonpointer==3.0.0
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.19
langchain-text-splitters==0.3.2
langsmith==0.1.143
libretranslatepy==2.1.1
lxml==5.3.0
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib-inline==0.1.7
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
numpy==1.26.4
openai==1.54.4
orjson==3.10.11
packaging==24.2
parso==0.8.4
prompt_toolkit==3.0.48
propcache==0.2.0
pure_eval==0.2.3
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
Pygments==2.18.0
python-docx==1.1.2
python-dotenv==1.0.1
pyvis==0.3.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
stack-data==0.6.3
tenacity==9.0.0
tqdm==4.67.0
traitlets==5.14.3
translate==3.6.1
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
wheel==0.44.0
yarl==1.17.1
